Batched Operations

NNlib implements a useful spectrum of operations designed to work with batches. At the moment, it is mainly a syntactic sugar when deployed on a CPU (as it doesn't seem to be hooked up to MKL); however, when running code on a GPU, it seamlessly integrates with CUDA, making use of the cuBLAS's native batched operations.

Though the most common use cases are covered by NNlib many aren't and it doesn't take long to encounter them. For now, this package implements some of the common batched operations encountered when working with IMU data that don't feature in NNlib; though only to the extent of being a syntactic sugar for operations on CPUs.

TODO

Once the imu.dev project reaches certain level of maturity we will aim to implement the most relevant batched operations on GPUs (by hooking to cuBLAS, which won't be difficult as CUDA.jl already does the heavy lifting), but for the time being, no concrete plans are made as to the exact date when that would happen.

Matrix-vector multiplication

IMUDevNNLib.batched_matvecmul — Function

alt_batched_mul(A::Abstract3Tensor, B::AbstractMatrix)

NNlib's batched_mul may not behave as expected when one of the arguments is a matrix. NNlib assumes that the matrix B must be broadcasted over each batch, however, what might instead be desired is to treat B as a batch of vectors for which a batch matrix-vector multiplication is desired.

Matrix inverse

IMUDevNNLib.batched_rinv — Function

batched_rinv(A::Abstract3Tensor)

Compute the inverse of each matrix in the batch A. Assumes that the last dimension is the batch dimension.

IMUDevNNLib.batched_rinvsolve — Function

batched_rinvsolve(A::Abstract3Tensor, B::Abstract3Tensor)

Compute the solution to the equation A⁽ⁱ⁾X = B⁽ⁱ⁾ for each matrix pair (A⁽ⁱ⁾, B⁽ⁱ⁾) in the batch. Assumes that the last dimension is the batch dimension.

IMUDevNNLib.batched_rinvsolve! — Function

batched_rinvsolve!(out::Abstract3Tensor, A::Abstract3Tensor, B::Abstract3Tensor)

Compute the solution to the equation A⁽ⁱ⁾X = B⁽ⁱ⁾ for each matrix pair (A⁽ⁱ⁾, B⁽ⁱ⁾) in the batch and save the results to out. Assumes that the last dimension is the batch dimension.