Data Layouts

DataLayouts define conventions for storing data.

Note

In imu.dev we are mostly interested in IMU data and these almost exclisively take the form of timeseries; however, in principle, any type of data can be represented with a suitable subtype of DataLayout.

IMUDevNNLib.DataLayout — Type

Parent to all defined data layouts. First and foremost, DataLayouts are merely conventions for consistently arranging data for analysis. From a more practical standpoint however, there are a couple of Base.reshape implementations that allow for transforming between different layouts. Additionally, there are also some utility functions.

Implemented layouts

IMUDevNNLib.SingleTimeseriesLayout — Type

Data stored in SingleTimeseriesLayout comprises of a single long timeseries stored in a single array. It's a (N-1)-dimensional array with

the first (N-2)-dimensions corresponding to the state space

the last (N-1)-th dimension corresponding to time

Note

There is only one sample, so one dimension that's usually reserved for samples (i.e. N-th dimension) is dropped.

Example

Consider a single recording from an IMU device that's been recording for one day @1Hz. The data extracted from this device represented in a Matrix form with dimensions 3×86_400, with each column representing a single observation of the 3d-acceleration would correspond to SingleTimeseriesLayout.

IMUDevNNLib.TimeseriesLayout — Type

Data stored in TimeseriesLayout comprises of a collection (usually a vector) of arrays, which conceptually represents a timeseries of observations for multiple samples. Each element of the collection is an (N-1)-dimensional array with:

the first (N-2)-dimensions corresponding to the state space

the last (N-1)-th dimension corresponding to samples

The collection indexes over timepoints (which can be thought of as the N-th dimension).

Example

Consider two recordings from two synchronized IMU devices that have been recording for one day @1Hz. The data extracted from these two devices represented in a Vector{Matrix} form with length 86_400 and each element of this vector having dimensions 3×2 (i.e. it would be two vectors of 3d-accelerations horizontally concatenated together into a 3×2 Matrix) would correspond to TimeseriesLayout.

IMUDevNNLib.SingleObsTimeseriesLayout — Type

Data stored as SingleObsTimeseriesLayout is simply a single array element (conceptually an observation) from the collection that's stored in TimeseriesLayout.

IMUDevNNLib.SingleArrayLayout — Type

Data stored in the SingleArrayLayout is an N-dimensional array with the last, N-th dimension indexing over samples.

Note

For a particular case of SingleArrayLayout intended to respresent time series data we follow the convention of using an N-dimensional array with:

the first (N-2)-dimensions corresponding to the state space

the (N-1)-th dimension corresponding to time, and

the N-th dimension corresponding to samples.

Utility functions

As DataLayouts impose some conventions about how the data are organized, knowing the type of DataLayout is sometimes already enough to extract useful information without needing to know any further specifics. We implement a couple of such utility functions:

IMUDevNNLib.state_dim — Function

state_dim(::Union{TimeseriesLayout,SingleObsTimeseriesLayout,
                  SingleTimeseriesLayout},
          data::AbstractArray)
          
state_dim(layout::TimeseriesLayout, data::Vector{<:AbstractArray})

Return the dimension of the state space of the time series data.

Note

The state space may in principle be represented by any AbstractArrays (i.e. not necessarily a Vector). Following the convention for Base.size the dimension is given as a tuple of integers.

IMUDevNNLib.num_samples — Function

num_samples(layout::DataLayout, data::AbstractArray)

num_samples(layout::TimeseriesLayout, data::Vector{<:AbstractArray})

num_samples(layout::SingleTimeseriesLayout, data::AbstractArray)

Return the total number of samples in data.

IMUDevNNLib.num_timepoints — Method

num_timepoints(layout::TimeseriesLayout, data::Vector{<:AbstractArray})

num_timepoints(layout::SingleArrayLayout, data::AbstractArray)

num_timepoints(layout::SingleTimeseriesLayout, data::AbstractArray)

Return the number of time points in the time series data.

Additionally, there are some further utility functions that make sense only under specific DataLayouts:

IMUDevNNLib.write_header! — Function

write_header!(io::IOStream, ::SingleArrayLayout, dims)

Utility function for writing the header of a binary file that stores data in SingleArrayLayout. dims is a tuple of integers representing the dimensions of the data.

IMUDevNNLib.read_header — Function

read_header(io::IOStream, ::SingleArrayLayout)

Utility function for reading the header of a binary file that stores data in SingleArrayLayout.

Finally, we provide routines to convert between different layouts:

Base.reshape — Function

Base.reshape(data::AbstractArray, from_to::Pair{LayoutA, LayoutB})

Transform data stored in LayoutA() into a format following LayoutB().

IMUDevNNLib.dimsofreshape — Function

dimsofreshape(data::Vector{<:BatchDelimitedArray},
              from_to::Pair{TimeseriesLayout,SingleArrayLayout})

Compute the size of the Array that would be needed to reshape data stored in TimeseriesLayout() into a format following SingleArrayLayout().

Tip

Using this function, as opposed to calling Base.reshape might be handy if it's not possible to perform the reshaping in memory, and instead, an Mmap array needs to be constructed.