Data transformations

Data transformation are very commonly employed in Machine Learning as a way to augment the training data.

IMUDevNNLib.DataTransformation — Type

abstract type DataTransformation

Parent to all preprocessing data transformations

Callable functions

(transf::DataTransformation)([rng::Random.AbstractRNG], data::AbstractArray...)

Apply transformation transf to the data.

Implemented data transformations

IMUDevNNLib.IdentityTransformation — Type

struct IdentityTransformation <: DataTransformation

Identity transformation. Does nothing to the data that's passed through it.

Callable functions

(::IdentityTransformation)([::Random.AbstractRNG],
                           ::DataTransformation,
                           data::AbstractArray...)

Pass data through the transformation. Nothing is done to it.

IMUDevNNLib.RandomPlanarRotation — Type

struct RandomPlanarRotation{T} <: DataTransformation

Random rotation of the data in the plane spanned by indices. The angle is chosen uniformly from the interval [-max_angle, max_angle].

Callable functions

(transf::RandomPlanarRotation)([rng::Random.AbstractRNG],
                               ::SingleObsTimeseriesLayout,
                               data::AbstractMatrix...)

Each matrix data is rotated by the same angle.

(transf::RandomPlanarRotation)(rng::Random.AbstractRNG,
                               ::StackedArrayLayout,
                               data::Abstract3Tensor...)

Each sample in data is rotated by a different angle, note however, that each tensor data is rotated by the same sequence of angles. In other words, sample i from each tensor data is rotated by the same angle and sample j from each tensor data is rotated by the same angle, but angles for samples i and j (with i≠j) are (in general) different from one another.

IMUDevNNLib.RandomShift — Type

struct RandomShift <: DataTransformation

Shift the data in the time dimension by a random amount chosen uniformly from the interval (-max_shift, max_shift).

Constructor

RandomShift(max_shift::Int)

Callable functions

(transf::RandomShift)(rng::Random.AbstractRNG,
                      ::StackedArrayLayout,
                      data::AbstractArray...)

It assumes that the data passed through it is a time series with the time dimension encompassing segment:

left_padding | window | right_padding

where left_padding and right_padding are the same and equal to max_shift.

RandomPlanarRotation will call function plane_rotation under the hood:

IMUDevNNLib.plane_rotation — Function

plane_rotation(data::AbstractMatrix, angle; plane_indices=1:2)

Rotate the data in the plane spanned by indices plane_indices by angle. The data are assumed to be given in a matrix form, with:

first dimension indexing through the state coordinates, and

the second dimension indexing through the time points

plane_rotation(data::Abstract3Tensor, angles::AbstractVector; plane_indices=1:2)

Rotate the data in the plane spanned by indices plane_indices by angles. The data are assumed to be given in a 3-tensor form, with:

first dimension indexing through the state coordinates,

the second dimension indexing through the samples, and

the third dimension indexing through the time points