Other

Some additional tools include:

IMUDevNNLib.skipfirstobs — Function

skipfirstobs(x::AbstractArray; timedim=ndims(x)-1)

Given an array x representing a time series, with the timedim dimension corresponding to time, return a new array with the first observation removed.

IMUDevNNLib.selectfirstobs — Function

selectfirstobs(x::AbstractArray; timedim=ndims(x)-1)

Given an array x representing a time series, with the timedim dimension corresponding to time, return a new array with only the first observation.

IMUDevNNLib.peelfirstobs — Function

peelfirstobs(x::AbstractArray; timedim=ndims(x)-1)

Given an array x representing a time series, with the timedim dimension corresponding to time, return two array views:

view into the first observation,

view into the remaining observations.

IMUDevNNLib.selectalonglastdim — Function

selectalonglastdim(m::AbstractArray, i)

Select subview of the array m with index i accessed along the last dimension of this array.

IMUDevNNLib.mergealonglastdim — Function

mergealonglastdim(data::Vector{<:AbstractArray})

Merge a collection of arrays along their last dimension. The arrays must agree in all dimensions except the last one.

Tip

yes, cat(data...; dims=ndims(first(data))) will work, but for large arrays it will be very, very inefficient, in which case you're better off using mergealonglastdim.

IMUDevNNLib.nnflatten — Function

nnflatten(x::AbstractArray{<:Any,N}) where {N}

A common flattening scheme encountered in machine learning libraries: flatten the array x into a 2D array by squashing all dimensions except the last one (i.e. except the batch dimension).