Temporal Data

Passing a time-series data through a Neural Network often requires a choice of an architecture that falls under the category of the so-called Recurrent Neural Networks (RNNs). Flux package (that the project imu.dev predominantly relies upon for the task of training Neural Nets) requires these data to adhere to a certain format. For this purpose we define a TemporalData struct to provide a unified way of handling training/validation/testing data to be passed through RNNs.

IMUDevNNLib.TemporalData — Type

TemporalData(xx::Vector{S}, yy::Vector{T}, x₀::S, y₀::T)

A type for temporal data, i.e. data that are indexed by time. Conceptually, xx and yy are "inputs" and "targets" respectively. Both xx and yy are vectors of arrays, where each array corresponds to a timepoint. Optionally, x₀ and y₀ can be provided, which correspond to the initial state and the initial observation (the latter is usually ignored).

Tip

Sometimes only one pair (x₀, xx) or (y₀, yy) is needed or available (for instance, in production during forecasting). In such cases, the other pair is usually left empty.

Constructors

temporal_data([T::DataType];
              xx::Union{<:AbstractArray,Missing}=missing,
              yy::Union{<:AbstractArray,Missing}=missing,
              x₀::Union{<:AbstractArray,Missing}=missing,
              y₀::Union{<:AbstractArray,Missing}=missing,
              skipfirst=false)

temporal_data([T::DataType], xx::AbstractArray, yy::AbstractArray;
              x₀::Union{<:AbstractArray,Missing}=missing,
              y₀::Union{<:AbstractArray,Missing}=missing,
              skipfirst=false)

Convenience constructor for TemporalData which accepts the data stored in single arrays (as opposed to vector of arrays). It expects that the last dimension of xx and yy correspond to the time dimension, whereas the penultimate dimension corresponds to the batch dimension. If T is provided, it will be used to convert the data to the specified type. If skipfirst is true, the first observation will be skipped for saving into internal xx and yy and stored separately as x₀ and y₀ (it is convenient, as it is not uncommon to treat the "zero'th" observation different from the rest). x₀ and y₀ may also be provided explicitly, in which case skipfirst will still skip the first observation, but it will be the user-provided x₀ and/or y₀ that will take the priority of being stored in the internal state of TemporalData as the inital states. At least one of xx or yy must be provided. If the other is left out, an empty placeholder will be created in its place.

Data Loader

The TemporalData implements the interface required by DataLoader, which means it can be passed to it directly

MLUtils.numobs — Function

MLUtils.numobs(d::TemporalData)

For temporal data, the number of observations is synonymous with the batch size. Batch size is going to be the last dimension of every array corresponding to any timepoint.

MLUtils.getobs — Function

MLUtils.getobs(d::TemporalData, i)

Return the entire time series for the i-th observation (if i is a vector or a range it will return a vector of time series for the correspoding batch.)

Example

data = ...
td = temporal_data(Float32,
                   data.train.xx,
                   data.train.yy; skipfirst=true)

loader = Flux.DataLoader(td; batchsize=20, shuffle=true)

Utility functions

Utility functions include

IMUDevNNLib.num_samples — Function

num_samples(td::TemporalData)

The number of availble samples in the temporal data (i.e. the maximal batch size).

IMUDevNNLib.num_timepoints — Function

num_timepoints(td::TemporalData)

The number of timepoints in the temporal data.

IMUDevNNLib.state_dim — Function

state_dim(td::TemporalData)

Return the dimension of the state space.

Note

The state dimension is known only if either the xx or x₀ are available. If both are missing, an error will be thrown.

IMUDevNNLib.obs_dim — Function

obs_dim(td::TemporalData)

Return the dimension of the observation space.

Note

The observation dimension is known only if either the yy or y₀ are available. If both are missing, an error will be thrown.

Base.size — Method

Base.size(td::TemporalData, of_what::Symbol; kwargs...)

Return the size of the output container for the given TemporalData object. This can be either:

the size of the state (if of_what is :state or :xx)

the size of the observation (if of_what is :obs, :observation or :yy)

Note that the output container is assumed to be an N-tensor (N > 2) where the last two dimensions correspond to the batch and time dimensions respectively; and in particular, the time dimension includes the initial state!

BatchedTemporalContainer

Even though TemporalData is the preferred format for working with RNNs, in the end, after we've run our forecasting algorithm we often want to come back to the usual format of one big N+2 dimensional matrix (N > 0), where N is the dimension of the state (usually 1 as we work with vectors) and where the last two dimensions are the batch (or samples) and time dimensions respectively. For this reason we define a BatchedTemporalContainer that facilitates easy writing of such data already at thetime of running our forecasting/testing algorithms.

IMUDevNNLib.BatchedTemporalContainer — Type

BatchedTemporalContainer(T::DataType,
                         size::NTuple{N,Int},
                         batchsize::Int) where {N}

A container that is used to write the results of operations performed on a TemporalData object. Internally, it is an N-tensor (N > 2) where the last two dimensions correspond to the batch and time dimensions respectively. Double indexing into this object via [i, t] returns the (N-2)-dimensional array corresponding to the t-th timepoint of the i-th batch. Single indexing [i] returns the entire time series of the i-th batch.

The container can be allocated directly from the TemporalData object by calling:

IMUDevNNLib.outputcontainer — Function

outputcontainer(td::TemporalData, for_what::Symbol; batchsize, kwargs...)

Create an output container for the given TemporalData object into which the output of smoothing can be conveniently saved in a batched manner. See BatchedTemporalContainer for more details. The for_what argument can be either :state or :obs (or any of their synonyms) to specify whether the output will be saved to xx or yy-sized container. The batchsize must correspond to the batch size at which the smoothing (or any other operation) is performed, so that indexing into the output container can be properly sized. obs_size or state_size can be specified to override the size of the output container. If not provided, the size will be inferred from the TemporalData (if possible).

Important

If td splits the initial state from the internal containers xx and yy, the output container will be enriched with the additional space to hold the initial state.

And the data can be retrieved with:

IMUDevNNLib.getdata — Function

getdata(c::BatchedTemporalContainer)

Return the underlying data of the container.

Example

Here is an example of how to use it:

model = ...
function smooth(loader::Flux.DataLoade{<:TemporalData})
    out = out_container(loader.data, :xx; batchsize=loader.batchsize)

    for (i, (x₀, _, _, yy)) in enumerate(loader)
        Flux.reset!(model)
        init!(model, x₀)
        # batch `i` timepoint `1`
        out[i, 1] .= x₀
        for (t, y) in enumerate(yy)
            x̂ₜ = model(y)
            # batch `i` timepoint `t+1`
            out[i, t + 1] = x̂ₜ
        end
    end

    return getdata(out)
end