Writing output

Once the training and testing is done and we have a model we wish to use in production we often need to save the results. The prediction happening in production stage may still happen in batches, but we often wish to save the data in a common SingleArrayLayout. To this end, we can make use of a utility struct BatchDelimitedArray to save to and read from using iteration that iterates over batches:

IMUDevNNLib.BatchDelimitedArray — Type

BatchDelimitedArray(data::AbstractArray, batchsize::Int)

An array data that is split into blocks of batches of size batchsize each (with the last batch possibly shrunk to fit the data's dimensions). Indexing effectively works like indexing over a Vector of subarrays, where each subarray corresponds to a batch.

Constructors

batchdelimitedarray(K::DataType, size::NTuple{<:Any,Int}, batchsize::Int)

Create a BatchDelimitedArray with underlying data of size and eltype K.

batchdelimitedarray(data::AbstractArray, batchsize::Int)

Same as the default BatchDelimitedArray's constructor.

Tip

Using a constructor that expects data::AbstractArray can be handy if you wish to use a Mmap-ed array.

Rather than accessing data through:

a::BatchDelimitedArray
a.data

it is recommended to use getdata function instead:

IMUDevNNLib.getdata — Function

getdata(a::BatchDelimitedArray)

Return the underlying data of the container.